apply CR
